# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/profiles/me": {
      "get": {
        "operationId": "ProfilesController_getMyProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns the current user profile."
          },
          "404": {
            "description": "Profile not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "profiles"
        ]
      },
      "post": {
        "operationId": "ProfilesController_createMyProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Profile created."
          },
          "400": {
            "description": "Underage."
          },
          "409": {
            "description": "Profile already exists."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "profiles"
        ]
      },
      "patch": {
        "operationId": "ProfilesController_updateMyProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "profiles"
        ]
      }
    },
    "/photos/upload-url": {
      "post": {
        "operationId": "PhotosController_getUploadUrl",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestUploadUrlDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Signed upload URL generated."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "photos"
        ]
      }
    },
    "/photos": {
      "post": {
        "operationId": "PhotosController_registerPhoto",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPhotoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Photo registered."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "photos"
        ]
      }
    },
    "/photos/{id}": {
      "get": {
        "operationId": "PhotosController_getPhoto",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a signed URL (blurred or original)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "photos"
        ]
      }
    },
    "/photos/{id}/reveal-grants": {
      "post": {
        "operationId": "PhotosController_grantReveal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantRevealDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reveal grant created."
          },
          "403": {
            "description": "Not the photo owner."
          },
          "409": {
            "description": "Grant already exists."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "photos"
        ]
      }
    },
    "/photos/{id}/reveal-grants/{viewerUserId}": {
      "delete": {
        "operationId": "PhotosController_revokeReveal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "viewerUserId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Grant revoked."
          },
          "403": {
            "description": "Not the photo owner."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "photos"
        ]
      }
    },
...
    "/blocks": {
      "post": {
        "operationId": "SafetyController_blockUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Block a user",
        "tags": [
          "safety"
        ]
      }
    },
    "/blocks/{id}": {
      "delete": {
        "operationId": "SafetyController_unblockUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Unblock a user",
        "tags": [
          "safety"
        ]
      }
    },
...
    "/resources": {
      "get": {
        "operationId": "ResourcesController_getPublishedResources",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Resources"
        ]
      },
      "post": {
        "operationId": "ResourcesController_createResource",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResourceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Resources"
        ]
      }
    },
    "/resources/{id}": {
      "patch": {
        "operationId": "ResourcesController_updateResource",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResourceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Resources"
        ]
      }
    },
    "/qa/threads": {
      "get": {
        "operationId": "QaController_getThreads",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Qa"
        ]
      },
      "post": {
        "operationId": "QaController_createThread",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQaThreadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Qa"
        ]
      }
    },
    "/qa/threads/{id}/reply": {
      "post": {
        "operationId": "QaController_replyToThread",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplyQaThreadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Qa"
        ]
      }
    },
    "/qa/threads/{id}/assign": {
      "post": {
        "operationId": "QaController_assignThread",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Qa"
        ]
      }
    },
    "/subscriptions/webhook/{provider}": {
      "post": {
        "operationId": "SubscriptionsController_handleWebhook",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Subscriptions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lib le Lib API",
    "description": "API documentation for the Lib le Lib application.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateProfileDto": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string",
            "example": "Alex"
          },
          "dateOfBirth": {
            "type": "string",
            "example": "1995-10-15"
          },
          "gender": {
            "type": "string",
            "enum": [
              "man",
              "woman",
              "other"
            ]
          },
          "regionId": {
            "type": "string",
            "description": "UUID of the region"
          },
          "relationshipGoals": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "marriage",
                "serious_relationship",
                "friendship"
              ]
            }
          },
          "bio": {
            "type": "string"
          },
          "discreetMode": {
            "type": "boolean"
          },
          "lowBandwidthMode": {
            "type": "boolean"
          },
          "preferredLanguage": {
            "type": "string",
            "enum": [
              "am",
              "en"
            ]
          },
          "interestTagIds": {
            "description": "Array of InterestTag UUIDs",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "nickname",
          "dateOfBirth",
          "gender",
          "regionId",
          "relationshipGoals"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string",
            "example": "Alex"
          },
          "dateOfBirth": {
            "type": "string",
            "example": "1995-10-15"
          },
          "gender": {
            "type": "string",
            "enum": [
              "man",
              "woman",
              "other"
            ]
          },
          "regionId": {
            "type": "string",
            "description": "UUID of the region"
          },
          "relationshipGoals": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "marriage",
                "serious_relationship",
                "friendship"
              ]
            }
          },
          "bio": {
            "type": "string"
          },
          "discreetMode": {
            "type": "boolean"
          },
          "lowBandwidthMode": {
            "type": "boolean"
          },
          "preferredLanguage": {
            "type": "string",
            "enum": [
              "am",
              "en"
            ]
          },
          "interestTagIds": {
            "description": "Array of InterestTag UUIDs",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "RequestUploadUrlDto": {
        "type": "object",
        "properties": {
          "contentType": {
            "type": "string",
            "example": "image/jpeg"
          }
        }
      },
      "RegisterPhotoDto": {
        "type": "object",
        "properties": {
          "storageRef": {
            "type": "string",
            "description": "The storage_ref key returned by /photos/upload-url"
          },
          "position": {
            "type": "number",
            "example": 0
          },
          "isPrimary": {
            "type": "boolean"
          }
        },
        "required": [
          "storageRef"
        ]
      },
      "GrantRevealDto": {
        "type": "object",
        "properties": {
          "viewerUserId": {
            "type": "string",
            "description": "UUID of the user to grant reveal access to"
          }
        },
        "required": [
          "viewerUserId"
        ]
      },
      "BlockUserDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateResourceDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateResourceDto": {
        "type": "object",
        "properties": {}
      },
      "CreateQaThreadDto": {
        "type": "object",
        "properties": {}
      },
      "ReplyQaThreadDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}